[refactor #50] PostCommandService 의존성 리팩토링
