+++ src/main/java/com/freepath/devpath/board/post/command/controller/PostCommandController.java
@@ -1,13 +1,13 @@
 package com.freepath.devpath.board.post.command.controller;
 
-import com.freepath.devpath.board.post.command.dto.PostCreateRequest;
-import com.freepath.devpath.board.post.command.dto.PostUpdateRequest;
+import com.freepath.devpath.board.post.command.dto.request.PostCreateRequest;
+import com.freepath.devpath.board.post.command.dto.request.PostUpdateRequest;
 import com.freepath.devpath.board.post.command.exception.FileDeleteFailedException;
 import com.freepath.devpath.board.post.command.exception.FileUpdateFailedException;
 import com.freepath.devpath.board.post.command.exception.InvalidPostAuthorException;
 import com.freepath.devpath.board.post.command.exception.NoSuchPostException;
 import com.freepath.devpath.board.post.command.service.PostCommandService;
-import com.freepath.devpath.board.post.command.dto.PostCreateResponse;
+import com.freepath.devpath.board.post.command.dto.response.PostCreateResponse;
 import com.freepath.devpath.common.exception.ErrorCode;
 import com.freepath.devpath.common.dto.ApiResponse;
 import lombok.RequiredArgsConstructor;
@@ -31,7 +31,7 @@ public class PostCommandController {
 
     @PostMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public ResponseEntity<ApiResponse<PostCreateResponse>> uploadPost(
-            @RequestPart PostCreateRequest postCreateRequest,
+            @RequestPart @Validated PostCreateRequest postCreateRequest,
             @RequestPart(value = "files", required = false) List<MultipartFile> multipartFiles,
             @AuthenticationPrincipal UserDetails userDetails
     ) {
